Pass native amount to estimate gas if non-zero
There's an upstream geth bug where gas estimation fails unless a
non-zero Value provided as a call arg. This bug doesn't affect Optimism
right now, but it will in the future once we upgrade to Bedrock. We pass
the Value now in rosetta so we don't regress after the Bedrock upgrade.
Providing the Value call arg is a no-op right now.

fixes #10