Support ERC20Votes delegate in ctor API
Metadata used to constuct delegate transactions on Optimism can be retrieved by requesting
/construction/metdata with the following intent:
```
[
  {
    "operation_identifier": {"index": 0},
    "type": "DELEGATE_VOTES",
    "account": {
      "address": <sender>
    },
    "amount": {
      "value": "0",
      "currency": {
        "symbol": "OP",
        "decimals": 18,
        "metadata": {
          "token_address": <token address>
        }
      }
    }
  },
  {
    "operation_identifier": {"index": 1},
    "type": "DELEGATE_VOTES",
    "account": {
      "address": <delegatee>
    },
    "amount": {
      "value": "0",
      "currency": {
        "symbol": "OP",
        "decimals": 18,
        "metadata": {
          "token_address": <token address>
        }
      }
    }
  }
]
```

An intent will be accepted as a delegation operation if it satisfies
these conditions:
1. There must be exactly two operations; the first representing an
account that's delegating votes, and the second represents the account
that's being delegated votes to.
2. The type of each operation must be set to "DELEGATE_VOTES".
3. The value of each operation must be set to zero.
4. The currency must contain a valid `token_address` in its metadata.
The token address should point to a [ERC20Votes](https://docs.openzeppelin.com/contracts/4.x/api/token/erc20#ERC20Votes) compatible token.

Once the metadata is created, it can be forwarded through the usual
flow to construct, sign and send a transaction on L2.