Fix crashing due to SELFDESTRUCT handling
This fixes a bug whereby an account commiting harakiri while making
itself the recipient causes its balance to be negative if it later sends
funds to another account (all within the same tx).

The OVM behaves differently from the EVM as a result of fixing the
saurik bug (https://www.saurik.com/optimism.html). The OVM models
SELFDESTRUCT as two operations; a credit to the SELFDESTRUCT recipient,
and a debit to the destroyed contract. The OVM, unlike the EVM, permits destroyed
accounts to retain its balance and thus send funds until the end of the
transaction.

Thus, this semantics of SELFDESTRUCT is tweaked to generate two
operations for the credit/debit transfer.