ctor: Fee estimation changes
The fee adjustment due to an out of date base fee is increased to reduce the changes of stuck
transactions in the mempool.
The gas fee cap is adjusted to quadruple the base fee.
The gas tip cap is reduced, while increasing the gas fee cap, to reduce the cost of replacing
transactions in the mempool.