Support the Ecotone upgrade
A fork-agnostic header type, `rpcHeader`, is introduced to ensure that
optimism-rosetta supports future network upgrades without manually
updating the header. In particular, we now rely on the block hash
provided via JSON-RPC, rather than computing it locally for
forwards-compatibility.

Though this is more of workaround the fact that we can't update the
op-geth dependency right now. This is because recent op-geth versions
pull in a version of gnark-crpyto that conflicts with rosetta-sdk-go.